requirements.txt
python-dotenv
discord.py
gitpython